DRY metadata forms
