+++ geonode/documents/forms.py
@@ -1,6 +1,5 @@
 import json
 import os
-import taggit
 import re
 
 from django import forms
@@ -10,68 +9,15 @@ from django.conf import settings
 from django.forms import HiddenInput, TextInput
 from modeltranslation.forms import TranslationModelForm
 
-from mptt.forms import TreeNodeMultipleChoiceField
-from bootstrap3_datetime.widgets import DateTimePicker
-
-from geonode.people.models import Profile
 from geonode.documents.models import Document
 from geonode.maps.models import Map
 from geonode.layers.models import Layer
-from geonode.base.models import Region
+from geonode.base.forms import ResourceBaseForm
 
 
-class DocumentForm(TranslationModelForm):
-    _date_widget_options = {
-        "icon_attrs": {"class": "fa fa-calendar"},
-        "attrs": {"class": "form-control input-sm"},
-        "format": "%Y-%m-%d %H:%M",
-        # Options for the datetimepickers are not set here on purpose.
-        # They are set in the metadata_form_js.html template because
-        # bootstrap-datetimepicker uses jquery for its initialization
-        # and we need to ensure it is available before trying to
-        # instantiate a new datetimepicker. This could probably be improved.
-        "options": False,
-    }
-    date = forms.DateTimeField(
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
-    temporal_extent_start = forms.DateTimeField(
-        required=False,
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
-    temporal_extent_end = forms.DateTimeField(
-        required=False,
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
+class DocumentForm(ResourceBaseForm):
 
     resource = forms.ChoiceField(label='Link to')
-
-    poc = forms.ModelChoiceField(
-        empty_label="Person outside GeoNode (fill form)",
-        label="Point Of Contact",
-        required=False,
-        queryset=Profile.objects.exclude(
-            username='AnonymousUser'))
-
-    metadata_author = forms.ModelChoiceField(
-        empty_label="Person outside GeoNode (fill form)",
-        label="Metadata Author",
-        required=False,
-        queryset=Profile.objects.exclude(
-            username='AnonymousUser'))
-
-    keywords = taggit.forms.TagField(
-        required=False,
-        help_text=_("A space or comma-separated list of keywords"))
-
-    regions = TreeNodeMultipleChoiceField(
-        required=False,
-        queryset=Region.objects.all(),
-        level_indicator=u'___')
-    regions.widget.attrs = {"size": 20}
 
     def __init__(self, *args, **kwargs):
         super(DocumentForm, self).__init__(*args, **kwargs)
@@ -91,18 +37,6 @@ class DocumentForm(TranslationModelForm):
             self.fields['resource'].initial = 'type:%s-id:%s' % (
                 self.instance.content_type.id, self.instance.object_id)
 
-        for field in self.fields:
-            help_text = self.fields[field].help_text
-            self.fields[field].help_text = None
-            if help_text != '':
-                self.fields[field].widget.attrs.update(
-                    {
-                        'class': 'has-popover',
-                        'data-content': help_text,
-                        'data-placement': 'right',
-                        'data-container': 'body',
-                        'data-html': 'true'})
-
     def save(self, *args, **kwargs):
         contenttype_id = None
         contenttype = None
@@ -119,39 +53,14 @@ class DocumentForm(TranslationModelForm):
         self.instance.content_type = contenttype
         return super(DocumentForm, self).save(*args, **kwargs)
 
-    class Meta:
+    class Meta(ResourceBaseForm.Meta):
         model = Document
-        exclude = (
-            'uuid',
-            'contacts',
-            'workspace',
-            'store',
-            'name',
-            'uuid',
-            'storeType',
-            'typename',
-            'bbox_x0',
-            'bbox_x1',
-            'bbox_y0',
-            'bbox_y1',
-            'srid',
-            'category',
-            'csw_typename',
-            'csw_schema',
-            'csw_mdsource',
-            'csw_type',
-            'csw_wkt_geometry',
-            'metadata_uploaded',
-            'metadata_xml',
-            'csw_anytext',
+        exclude = ResourceBaseForm.Meta.exclude + (
             'content_type',
             'object_id',
             'doc_file',
             'extension',
             'doc_type',
-            'popular_count',
-            'share_count',
-            'thumbnail',
             'doc_url')
