+++ geonode/layers/forms.py
@@ -20,21 +20,13 @@
 
 import os
 import tempfile
-import taggit
+import autocomplete_light
 
 from django import forms
 from django.utils import simplejson as json
-from django.utils.translation import ugettext as _
-from modeltranslation.forms import TranslationModelForm
-
-from mptt.forms import TreeNodeMultipleChoiceField
-from bootstrap3_datetime.widgets import DateTimePicker
 
 from geonode.layers.models import Layer, Attribute
-from geonode.people.models import Profile
-from geonode.base.models import Region
-
-import autocomplete_light
+from geonode.base.forms import ResourceBaseForm
 
 
 class JSONField(forms.CharField):
@@ -47,107 +39,20 @@ class JSONField(forms.CharField):
             raise forms.ValidationError("this field must be valid JSON")
 
 
-class LayerForm(TranslationModelForm):
-    _date_widget_options = {
-        "icon_attrs": {"class": "fa fa-calendar"},
-        "attrs": {"class": "form-control input-sm"},
-        "format": "%Y-%m-%d %H:%M",
-        # Options for the datetimepickers are not set here on purpose.
-        # They are set in the metadata_form_js.html template because
-        # bootstrap-datetimepicker uses jquery for its initialization
-        # and we need to ensure it is available before trying to
-        # instantiate a new datetimepicker. This could probably be improved.
-        "options": False,
-        }
-    date = forms.DateTimeField(
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
-    temporal_extent_start = forms.DateTimeField(
-        required=False,
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
-    temporal_extent_end = forms.DateTimeField(
-        required=False,
-        localize=True,
-        widget=DateTimePicker(**_date_widget_options)
-    )
+class LayerForm(ResourceBaseForm):
 
-    poc = forms.ModelChoiceField(
-        empty_label="Person outside GeoNode (fill form)",
-        label="Point Of Contact",
-        required=False,
-        queryset=Profile.objects.exclude(
-            username='AnonymousUser'),
-        widget=autocomplete_light.ChoiceWidget('ProfileAutocomplete'))
-
-    metadata_author = forms.ModelChoiceField(
-        empty_label="Person outside GeoNode (fill form)",
-        label="Metadata Author",
-        required=False,
-        queryset=Profile.objects.exclude(
-            username='AnonymousUser'),
-        widget=autocomplete_light.ChoiceWidget('ProfileAutocomplete'))
-
-    keywords = taggit.forms.TagField(
-        required=False,
-        help_text=_("A space or comma-separated list of keywords"))
-
-    regions = TreeNodeMultipleChoiceField(
-        required=False,
-        queryset=Region.objects.all(),
-        level_indicator=u'___')
-    regions.widget.attrs = {"size": 20}
-
-    class Meta:
+    class Meta(ResourceBaseForm.Meta):
         model = Layer
-        exclude = (
-            'contacts',
+        exclude = ResourceBaseForm.Meta.exclude + (
             'workspace',
             'store',
-            'name',
-            'uuid',
             'storeType',
             'typename',
-            'bbox_x0',
-            'bbox_x1',
-            'bbox_y0',
-            'bbox_y1',
-            'srid',
-            'category',
-            'csw_typename',
-            'csw_schema',
-            'csw_mdsource',
-            'csw_type',
-            'csw_wkt_geometry',
-            'metadata_uploaded',
-            'metadata_xml',
-            'csw_anytext',
-            'popular_count',
-            'share_count',
-            'thumbnail',
             'default_style',
             'styles',
             'upload_session',
-            'charset',
-            'rating',
             'service',)
         widgets = autocomplete_light.get_widgets_dict(Layer)
-
-    def __init__(self, *args, **kwargs):
-        super(LayerForm, self).__init__(*args, **kwargs)
-        for field in self.fields:
-            help_text = self.fields[field].help_text
-            self.fields[field].help_text = None
-            if help_text != '':
-                self.fields[field].widget.attrs.update(
-                    {
-                        'class': 'has-popover',
-                        'data-content': help_text,
-                        'data-placement': 'right',
-                        'data-container': 'body',
-                        'data-html': 'true'})
 
 
 class LayerUploadForm(forms.Form):
